requirements.txt
astroid==2.0.4
beautifulsoup4==4.4.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.3
crayons==0.1.2
django>=2.0.11
django-ajax-selects==1.7.1
django-appconf==1.0.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.2.1
django-classy-tags==0.8.0
django-coverage==1.2.4
django-crispy-forms==1.7.2
mistune==0.8.4
django-filter
coverage
django-recurrence
bleach
python-coveralls
selenium
pyvirtualdisplay
